+++ api/package.json
@@ -79,6 +79,7 @@
     "commander": "^5.1.0",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
+    "csv-express": "^1.2.2",
     "dotenv": "^8.2.0",
     "execa": "^4.0.3",
     "exif-reader": "^1.0.3",
@@ -92,6 +93,7 @@
     "inquirer": "^7.3.3",
     "joi": "^17.1.1",
     "js-yaml": "^3.14.0",
+    "json2csv": "^5.0.1",
     "jsonwebtoken": "^8.5.1",
     "knex": "^0.21.4",
     "knex-schema-inspector": "0.0.9",
