+++ api/src/exceptions/index.ts
@@ -1,5 +1,6 @@
 export * from './base';
 export * from './collection-not-found';
+export * from './export-failed';
 export * from './field-not-found';
 export * from './forbidden';
 export * from './invalid-credentials';
